Refactoring - move some types to internal from shared
